+++ src/diffFiles.ts
@@ -25,7 +25,7 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
     this._onDidChangeTreeData.event;
 
   constructor(
-    private bookPath: string,
+    private bookPath?: string,
     private filesBaseSrc: DiffFileInfo[] = [],
     private filesBaseVbe: DiffFileInfo[] = []
   ) {}
@@ -34,7 +34,11 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
     this._onDidChangeTreeData.fire();
   }
 
-  setFiles(bookPath: string, filesBaseSrc: DiffFileInfo[], filesBaseVbe: DiffFileInfo[]) {
+  setFiles(
+    bookPath?: string,
+    filesBaseSrc: DiffFileInfo[] = [],
+    filesBaseVbe: DiffFileInfo[] = []
+  ) {
     this.bookPath = bookPath;
     this.filesBaseSrc = filesBaseSrc;
     this.filesBaseVbe = filesBaseVbe;
@@ -44,7 +48,7 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
     if (element.isHeader) {
       const treeItem = new FileDiff(
         element.moduleName ?? STRING_EMPTY,
-        vscode.TreeItemCollapsibleState.Collapsed
+        vscode.TreeItemCollapsibleState.Expanded
       );
       return treeItem;
     } else {
@@ -83,6 +87,7 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
   }
 
   async updateModification(bookPath?: string) {
+    const oldBookPath = this.bookPath;
     let bookPathToUpdate = bookPath;
     if (!bookPathToUpdate) {
       const activeTextEditor = vscode.window.activeTextEditor?.document.uri;
@@ -92,15 +97,22 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
     if (bookPathToUpdate && (await vbecmCommon.fileExists(bookPathToUpdate))) {
       // update path
       this.bookPath = bookPathToUpdate;
-      vbeTreeView.title = 'diff: ' + path.basename(bookPathToUpdate);
     }
 
     if (this.bookPath === undefined) {
+      throw Error('Select vba modules.');
       return;
+    }
+
+    if (oldBookPath !== this.bookPath) {
+      //refresh
+      // fileDiffProvider.setFiles(this.bookPath, [], []);
+      // fileDiffProvider.refresh();
     }
 
     try {
       const { diffBaseSrc, diffBaseVbe } = await compareModules(this.bookPath);
+      vbeTreeView.title = 'diff: ' + path.basename(this.bookPath);
       fileDiffProvider.setFiles(this.bookPath, diffBaseSrc, diffBaseVbe);
       fileDiffProvider.refresh();
     } catch (error) {
@@ -126,7 +138,7 @@ export class FileDiffProvider implements vscode.TreeDataProvider<DiffFileInfo> {
   }
 }
 
-export const fileDiffProvider = new FileDiffProvider(STRING_EMPTY, [], []);
+export const fileDiffProvider = new FileDiffProvider(undefined, [], []);
 export const vbeTreeView = vscode.window.createTreeView('vbeDiffView', {
   treeDataProvider: fileDiffProvider,
 });